Voice recognition based Safety System for a car
This project is based on hardware solution for voice recognition. The
project includes Voice recognition module, GSM module and GPS module.
The entire setup is placed on the vehicle. In case of vehicle theft,
user can access this module to get the position of vehicle. For
accessing this module, user calls on the GSM module attached to the
setup which is then connected to Voice recognition module. Voice
recognition module authenticates user voice and then allows the access.
On demanding the position, GPS module in the setup sends accurate
vehicle coordinates on the registered mobile. This is used in security
system. This project also helps in limiting the over speeding of vehicle
and controlling the headlight beam of vehicle.